Split product and homepage tests
